+++ Editor/FileUtility.cs
@@ -39,7 +39,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 			return path.Replace(string.Concat(WinSeparator, WinSeparator), WinSeparator.ToString());
 		}
 
-		public static string NormalizeWindowsToUnix(string path)
+		public static string NormalizeWindowsToUnix(this string path)
 		{
 			if (string.IsNullOrEmpty(path))
 				return path;